% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\docType{class}
\name{Background-class}
\alias{Background-class}
\alias{.Background}
\title{Background class definition}
\description{
Objects of this class serve as a container that holds
parameters for the Background model.
}
\details{
A Background model is constructed 
via \code{\link{readBackground}}.
}
\section{Slots}{

\describe{
\item{\code{station}}{Stationary probabilities}

\item{\code{trans}}{Transition probabilities}

\item{\code{counts}}{k-mer counts}

\item{\code{order}}{Background model order}
}}

