% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{backgroundValid}
\alias{backgroundValid}
\title{Check valididity of Background model}
\usage{
backgroundValid(bg)
}
\arguments{
\item{bg}{A Background object}
}
\value{
None
}
\description{
This function checks if the Background model is valid. The function throws
an error if the obejct does not represent a Background model.
}
\examples{

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)

# estimate background model from seqfile
bg=readBackground(seqs,1)
backgroundValid(bg)

}

