% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{bgOrderForSampling}
\alias{bgOrderForSampling}
\title{Get background order}
\usage{
bgOrderForSampling()
}
\value{
Order of the background model
}
\description{
This function returns the background model order
}

