% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{deleteBackgroundForSampling}
\alias{deleteBackgroundForSampling}
\title{Delete the current background model for sampling}
\usage{
deleteBackgroundForSampling()
}
\description{
This function unloads the current background model for sampling.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

library(mdist)
seqfile=system.file("extdata","seq.fasta", package="mdist")
readBackgroundForSampling(file=seqfile,1)
deleteBackgroundForSampling()


}

