% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{fetchBackground}
\alias{fetchBackground}
\title{Fetch background model into R vectors}
\usage{
fetchBackground()
}
\value{
A list containing a vector containing the stationary
             and the transition probabilities, respectively.
}
\description{
This function fetches the current stationary distribution
and transition probabilities into R two vectors that are
contained in list. This function is only used for debugging 
and testing purposes.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

library(mdist)
file=system.file("extdata","seq.fasta", package="mdist")
readBackground(file,1)
fetchBackground()

}

