% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_wrapper.R
\name{generateDNAString}
\alias{generateDNAString}
\title{Generate DNAString}
\usage{
generateDNAString(len)
}
\arguments{
\item{len}{Length of the sequence}
}
\value{
DNAString
}
\description{
This function generates a DNAString
by sampling from the the background model.
}
\examples{

# Set the the significance level and the score granularity
motifcounterOption(alpha=0.01, gran=0.1)

seqfile=system.file("extdata","seq.fasta", package="motifcounter")

# Load an order-1 background model
readBackgroundForSampling(seqfile,1)

# generates a sequences of length 1kb
generateDNAString(100)

}
\seealso{
\code{\link{readBackgroundForSampling}}
}

