% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forground_wrapper.R
\name{getPositionWeights}
\alias{getPositionWeights}
\title{Create position weight representation}
\usage{
getPositionWeights(pfm, bg)
}
\arguments{
\item{pfm}{An R matrix that represents a position frequency matrix}

\item{bg}{A Background object}
}
\value{
Position weight representation
}
\description{
This function computes the position weight representation for a given
PFM and a background model which corresponds to the log-likelihood
ratio for each position. For order=0, this corresponds to the
commonly used (4 x M) position weight matrix where M denotes
the motif length.
For higher order background, we obtain a pow(4, order+1) x (M - order + 1)
dimensional matrix.
}
\examples{

# Load sequences
seqfile = system.file("extdata", "seq.fasta", package = "motifcounter")
seqs = Biostrings::readDNAStringSet(seqfile)

# Load background
bg = readBackground(seqs, 1)

# Load motif
motiffile = system.file("extdata", "x1.tab", package = "motifcounter")
motif = t(as.matrix(read.table(motiffile)))

# Reverse complement motif
pwm = motifcounter:::getPositionWeights(motif, bg)

}
