% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{getSinglestranded}
\alias{getSinglestranded}
\title{Accessor to slot singlestranded}
\usage{
getSinglestranded(obj)
}
\arguments{
\item{obj}{An Overlap object}
}
\value{
singlestranded slot
}
\description{
Accessor to slot singlestranded
}
