% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{getStation}
\alias{getStation}
\title{Accessor to slot station}
\usage{
getStation(obj)
}
\arguments{
\item{obj}{A Background object}
}
\value{
station slot
}
\description{
Accessor to slot station
}

