% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_wrapper.R
\name{lenSequences}
\alias{lenSequences}
\title{Length of sequences in a given fasta file}
\usage{
lenSequences(seqs)
}
\arguments{
\item{seqs}{DNAStringSet}
}
\value{
A vector containing the lengths of each individual sequence in the
fasta file
}
\description{
The function parses a fasta file and returns a vector containing the length
of each individual sequence contained in the fasta file. Sequences
containing 'N' or 'n' are skipped from the analysis and are
set to length zero.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)

# scan the given sequence for the motif occurances
lseq=lenSequences(seqs)
lseq

}

