% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forground_wrapper.R
\name{motif2matrix}
\alias{motif2matrix}
\title{Get the motif in an R matrix}
\usage{
motif2matrix()
}
\description{
This function returns the current PFM as an R matrix.
}
\examples{


library(mdist)
pwmfile=system.file("extdata","x31.tab", package="mdist")
readMotif(pwmfile)
x=motif2matrix()

}

