% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{motifcounterOption}
\alias{motifcounterOption}
\title{Set parameters for the enrichment analysis}
\usage{
motifcounterOption(alpha = 0.001, gran = 0.1, ncores = 1)
}
\arguments{
\item{alpha}{Significance level for calling
motif hits. E.g. alpha=0.001 amounts to calling one motif hit per strand
by chance in a sequence of length 1000. Default: alpha=0.001}

\item{gran}{The score granularity which is used
for discretizing the score range. Decreasing gran values
will increase number
of discrete bins that represent the real-valued score range.
This will yield more a accurate score distribution due to less
discretization noise, however, it incurs an increase of the computational
runtime.  Default: gran=0.1}

\item{ncores}{Number of cores used for parallel processing if openMP is
available. Default: ncores=1.}
}
\value{
None
}
\description{
This function sets some parameters that are necessary for
the computations of the `motifcounter` package.
}
\examples{


motifcounterOption(alpha=0.001)

}

