% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentTest.R
\name{motifEnrichment_}
\alias{motifEnrichment_}
\title{Internal logic Enrichment of motif hits}
\usage{
motifEnrichment_(seqs, pfm, bg, singlestranded = FALSE,
  method = "compound", ...)
}
\arguments{
\item{seqs}{A DNAStringSet or DNAString object}

\item{pfm}{An R matrix that represents a position frequency matrix}

\item{bg}{A Background object}

\item{singlestranded}{Boolean that indicates whether a single strand or
both strands shall be scanned for motif hits.
Default: singlestranded = FALSE.}

\item{method}{String that defines whether to use
the 'compound' Poisson approximation', the 'combinatorial' model
or whether to skip P-value evaluation ('nopval').
Default: method='compound'.}

\item{...}{externally supplied overlap object}
}
\value{
List that contains
\describe{
\item{numofhits}{Number of observations}
\item{pvalue}{P-value for the enrichment test}
\item{fold}{Fold-enrichment with respect to the expected number of hits}
\item{logfold}{log-Fold-enrichment with respect to
               the expected number of hits}
}
}
\description{
This function determines whether a given motif is enriched in a given
DNA sequences.
}
\details{
Enrichment is tested by comparing the observed
number of motif hits against a theoretical distribution of the number
of motif hits in random DNA sequences.
Optionally, the theoretical distribution of the number of motif
hits can be evaluated by either a 'compound Poisson model'
or the 'combinatorial model'.
Additionally, the enrichment test can be conducted with respect
to scanning only the forward strand or both strands of the DNA
sequences. The latter option is only available for the
'compound Poisson model'.
}
\seealso{
\code{\link{compoundPoissonDist}}, \code{\link{combinatorialDist}}
}
