% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentTest.R
\name{motifEnrichmentTest}
\alias{motifEnrichmentTest}
\title{Enrichment of motif hits}
\usage{
motifEnrichmentTest(observations, overlap, method = "compound")
}
\arguments{
\item{observations}{List containing the number of motif hit 
in the given sequence. This argument is
determined by \code{\link{numMotifHits}}.}

\item{overlap}{List containing the overlapping hit probabilities. 
This argument is determined by  \code{\link{probOverlapHit}}.}

\item{method}{String that defines whether to use 
the 'compound' Poisson approximation' or the 'combinatorial' model.
Default: method='compound'}
}
\value{
P-value for the enrichment test
}
\description{
This function determines whether a given motif is enriched in a given 
DNA sequences based on the number of motif hits that would be observed
in a random sequence that is generated by the background model. 
It can optionally be used based on the 
'compound Poisson approximation' or the 'combinatorial model'.
}
\examples{


seqfile=system.file("extdata","seq.fasta", package="mdist")
motiffile=system.file("extdata","x31.tab", package="mdist")
alpha=0.001
gran=0.1
mdistOption(alpha, gran)

# estimate background model from seqfile
readBackground(seqfile,1)

# load motif model from motiffile
readMotif(motiffile, 0.01)

### 1 ) Compute the distribution for scanning a *single* DNA strand

# Compute overlapping probabilities 
op=probOverlapHit(singlestranded=TRUE)

# Determine the observed number of motif hits in 'seq.fasta'
numhits=numMotifHits(seqfile,singlestranded=TRUE)

# Determine the enrichment of the motif in the sequence
# based on the 'Compound Poisson model'

pvalue=motifEnrichmentTest(numhits,op,method="compound")

### 2 ) Compute the distribution for scanning *both* DNA strand

#Compute overlapping probabilities 
op=probOverlapHit(singlestranded=FALSE)

# Determine the observed number of motif hits in 'seq.fasta'
numhits=numMotifHits(seqfile,singlestranded=FALSE)

# Determine the enrichment of the motif in the sequence
# based on the 'Compound Poisson model'

pvalue=motifEnrichmentTest(numhits,op,method="compound")

### 3 ) Compute the distribution for scanning *both* DNA strand

#Compute overlapping probabilities 
op=probOverlapHit(singlestranded=FALSE)

#Determine the observed number of motif hits in 'seq.fasta'
numhits=numMotifHits(seqfile,singlestranded=FALSE)

# Determine the enrichment of the motif in the sequence
# based on the *combinatorial model*

pvalue=motifEnrichmentTest(numhits,op,method="combinatorial")

}
\seealso{
\code{\link{compoundPoissonDist}}, \code{\link{combinatorialDist}}

\code{\link{numMotifHits}}, \code{\link{probOverlapHit}}
}

