% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forground_wrapper.R
\name{motifLength}
\alias{motifLength}
\title{Length of the currently loaded PFM}
\usage{
motifLength()
}
\value{
Length of the motif
}
\description{
This function returns the length of the currently loaded PFM.
}
\examples{


motiffile=system.file("extdata","x31.tab", package="motifcounter")
readMotif(motiffile)
motifLength()

}

