% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forground_wrapper.R
\name{normalizeMotif}
\alias{normalizeMotif}
\title{Add pseudo-count to PFM}
\usage{
normalizeMotif(pfm, pseudo = 0.01)
}
\arguments{
\item{pfm}{A position frequency matrix}

\item{pseudo}{Small pseudo-value that is added to each entry in the PFM}
}
\value{
A renormalized pfm
}
\description{
This function can be used to prevent the PFM from
containing zero-value entries.
}
\examples{


motiffile=system.file("extdata","x1.tab", package="motifcounter")
motif=t(as.matrix(read.table(motiffile)))
new_motif=normalizeMotif(motif)

}

