% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_wrapper.R
\name{numMotifHits}
\alias{numMotifHits}
\title{Number of motif hits in a given DNA sequence}
\usage{
numMotifHits(seqs, pfm, bg, singlestranded = FALSE)
}
\arguments{
\item{seqs}{A DNAStringSet object}

\item{pfm}{An R matrix that represents a position frequency matrix}

\item{bg}{A Background object}

\item{singlestranded}{Boolean that indicates whether a single strand or
both strands shall be scanned for motif hits. 
Default: singlestranded = FALSE.}
}
\value{
A list containing
\describe{
\item{nseq}{Number of individual sequences}
\item{lseq}{Vector of individual sequence lengths}
\item{numofhits}{Vector of the number of hits in each individual sequence}
}
}
\description{
This function counts the number of motif hits that
are found in the supplied DNA sequences.
It can be used to count motif hits on 
one or both strands, respectively.
}
\examples{

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
motiffile=system.file("extdata","x31.tab", package="motifcounter")
seq=Biostrings::readDNAStringSet(seqfile)

# Set false positive probability
alpha=0.001
motifcounterOption(alpha)

# Estimate an order-1 background model
bg=readBackground(seq,1)
# read PFM from file
motif=t(as.matrix(read.table(motiffile)))

# scan the given sequence on both strands for the motif occurrences
noc=motifcounter:::numMotifHits(seq,motif,bg)
noc

# scan the given sequence on a single strand for the motif occurences
noc=motifcounter:::numMotifHits(seq,motif,bg,singlestranded=TRUE)
noc

}

