% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_wrapper.R
\name{numSequences}
\alias{numSequences}
\title{Number of sequence in a provided fasta file}
\usage{
numSequences(seqfile)
}
\arguments{
\item{seqfile}{Name of the sequence file in fasta format.}
}
\description{
The function parses a fasta file and returns the number of individual
sequence in it.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

library(mdist)
seqfile=system.file("extdata","seq.fasta", package="mdist")

# scan the given sequence for the motif occurances
nseq=numSequences(seqfile)
nseq

}

