% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\docType{class}
\name{Overlap-class}
\alias{.Overlap}
\alias{Overlap-class}
\title{Overlap class definition}
\description{
Objects of this class serve as a container that holds
parameters for the overlapping hit probabilities.
}
\details{
An Overlap object is constructed via the \code{\link{probOverlapHit}}
}
\section{Slots}{

\describe{
\item{\code{alpha}}{Scalar numeric significance level to call motif matches}

\item{\code{beta}}{Numeric vector of 
principal overlapping hit probabilities on the same strand.}

\item{\code{beta3p}}{Numeric vector of
principal overlapping hit probabilities with 3'-overlap.}

\item{\code{beta5p}}{Numeric vector of
principal overlapping hit probabilities with 5'-overlap.}

\item{\code{gamma}}{Numeric vector of marginal overlapping hit probabilities.}

\item{\code{singlestranded}}{logical flag to indicate whether one
or both strands are scanned for motif matches.}
}}

