% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlapValid}
\alias{overlapValid}
\title{Check validity of Overlap}
\usage{
overlapValid(overlap)
}
\arguments{
\item{overlap}{An Overlap object}
}
\value{
None
}
\description{
This function checks if the Overlap object is valid. The function throws
an error if the obejct does not represent a Overlap object.
}

