% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{readBackground}
\alias{readBackground}
\title{Estimates the background model from a set of DNA sequences}
\usage{
readBackground(seqs, order = 1)
}
\arguments{
\item{seqs}{DNAStringSet}

\item{order}{Order of the Markov models that shall be used as the
background model. Default: order=1.}
}
\value{
A Background object
}
\description{
Given a set of DNA sequences and an order, this function
estimates a higher-order Markov model which is used as a
background model.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

file=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(file)
bg=readBackground(seqs,1)

}

