% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forground_wrapper.R
\name{readMotif}
\alias{readMotif}
\title{Read and load the PFM}
\usage{
readMotif(pwm, pseudocount = 0.01)
}
\arguments{
\item{pwm}{Character string of the name of the file which contains the PFM
or an R matrix.}

\item{pseudocount}{Numeric value which is added to each element of the
loaded PFM in order to avoid zero probability elements. The columns of the
matrix are renormalized automatically.}
}
\description{
The function reads a position frequency matrix (PFM) for scanning a DNA
sequence. The PFM can be loaded directly in transfac format, in tab format
or from an R matrix. To load a PFM from a file directly, the files must end
in '.transfac' or '.tab' to load from transfac or tab format, respectively.
}
\examples{


library(mdist)
pwmfile=system.file("extdata","x31.tab", package="mdist")
# load a motif in tab format
readMotif(pwmfile, 0.01)

# fetch the motif to an R matrix
pwm=motif2matrix()

#reload the motif again from the R matrix
readMotif(pwm, 0.01)


}

