% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_wrapper.R
\name{scoreDist}
\alias{scoreDist}
\title{Score distribution}
\usage{
scoreDist(pfm, bg)
}
\arguments{
\item{pfm}{A position frequency matrix}

\item{bg}{A Background object}
}
\value{
List containing
\describe{
\item{score}{Vector of motif scores}
\item{probability}{Score probabilities}
}
}
\description{
This function computes the score distribution for the provided PFM and
background. The Score distribution is computed based on an efficient
dynamic programming algorithm.
}
\examples{


# Set the significance level and granularity for the score computation
motifcounterOption(alpha=0.01,gran=0.1)

# Load the DNA sequence and a Motif
seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)
motiffile=system.file("extdata","x31.tab",package="motifcounter")

# Load an order-1 background model
bg=readBackground(seqs,1)

# Load the motif from the motiffile
motif=t(as.matrix(read.table(motiffile)))

# Compute the score distribution
dp=scoreDist(motif,bg)

}

