% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_wrapper.R
\name{scoreDistEmpirical}
\alias{scoreDistEmpirical}
\title{Empirical score distribution}
\usage{
scoreDistEmpirical(pfm, bg, seqlen, nsim)
}
\arguments{
\item{pfm}{An R matrix that represents a position frequency matrix}

\item{bg}{A Background object}

\item{seqlen}{Integer-valued vector that defines the lengths of the
individual sequences. For a given DNAStringSet, 
this information can be retrieved using \code{\link{numMotifHits}}.}

\item{nsim}{Integer number of random samples.}
}
\value{
List containing
\describe{
\item{score}{Vector of scores}
\item{probability}{Vector of score probabilities}
}
}
\description{
This function estimates the empirical score distribution
by simulating random DNA sequences based on the
background model. Subsequently, the random sequences are scanned
with the scoring measure which yields the score distribution.
This function is only used for benchmarking analysis.
}
\examples{

# Set the the significance level and the score granularity
motifcounterOption(alpha=0.01, gran=0.1)

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)
motiffile=system.file("extdata","x31.tab",package="motifcounter")

# Load an order-1 background model
bg=readBackground(seqs,1)

# Load a motif
motif=t(as.matrix(read.table(motiffile)))

# generate the simulated score distribution on
# sequences of length 1kb using 1000 samples
motifcounter:::scoreDistEmpirical(motif,bg,seqlen= 1000,nsim=1000)

}
\seealso{
\code{\link{scoreDist}}
}

