% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_wrapper.R
\name{scoreHistogram}
\alias{scoreHistogram}
\title{Score histogram}
\usage{
scoreHistogram(seqs, pfm, bg)
}
\arguments{
\item{seqs}{A DNAStringSet object}

\item{pfm}{An R matrix that represents a position frequency matrix}

\item{bg}{A Background object}
}
\value{
List containing
\describe{
\item{score}{Vector of score bins}
\item{probability}{Empirical frequencies associated with each score}
}
}
\description{
This function computes the empirical score
distribution for a given DNAStringSet object.
}
\details{
It can be used to compare the empirical score
distribution against the theoretical one (see \code{\link{scoreDist}}).
}
\examples{

# Set the the significance level and the score granularity
motifcounterOption(alpha=0.01, gran=0.1)

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
motiffile=system.file("extdata","x31.tab",package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)

# Load an order-1 background model
bg=readBackground(seqs,1)

# Load a motif from the motiffile
motif=t(as.matrix(read.table(motiffile)))


# Compute the empirical score histogram
scoreHistogram(seqs,motif,bg)

}
\seealso{
\code{\link{scoreDist}}
}

