% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_wrapper.R
\name{scoreSequenceProfile}
\alias{scoreSequenceProfile}
\title{Score profile across multiple sequences}
\usage{
scoreSequenceProfile(seqs, pfm, bg)
}
\arguments{
\item{seqs}{A DNAStringSet object}

\item{pfm}{An R matrix that represents a position frequency matrix}

\item{bg}{A Background object}
}
\value{
List containing
\describe{
\item{fscores}{Vector of per-position average forward strand scores}
\item{rscores}{Vector of per-position average reverse strand scores}
}
}
\description{
This function computes the per-position and per-strand 
average score profiles across a set of DNA sequences.
}
\examples{



# Load the DNA sequence and a Motif
seqfile=system.file("extdata","oct4_chipseq.fa", package="motifcounter")
motiffile=system.file("extdata","x31.tab",package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)
seqs=seqs[1:10]

# Load an order-1 background model
bg=readBackground(seqs,1)

# Load a motif
motif=t(as.matrix(read.table(motiffile)))

# Compute the score profile
scoreSequenceProfile(seqs,motif,bg)

}

