% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{sigLevel}
\alias{sigLevel}
\title{Retrieve the false positive level}
\usage{
sigLevel()
}
\value{
False positive motif hit probability
}
\description{
This function returns the current false positive level for
calling motif hits in random sequences.
}
\examples{
motifcounter:::sigLevel()

}

