% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial_wrapper.R
\name{combinatorialDist}
\alias{combinatorialDist}
\title{Combinatrial model approximation of the number of motif hits}
\usage{
combinatorialDist(seqlen, overlap)
}
\arguments{
\item{seqlen}{Integer-valued vector that defines the lengths of the
individual sequences. For a given DNAStringSet, 
this information can be retrieved using \code{\link{numMotifHits}}.}

\item{overlap}{An Overlap object which is created by
\code{\link{probOverlapHit}}.}
}
\description{
This function approxmiates the distribution of the number of motif hits.
It sums over all combinations of obtaining k hits in a random sequence
of a given length using an efficient dynamic programming algorithm.
}
\details{
This function is an alternative to \code{\link{compoundPoissonDist}}.
However, it requires fixed-length sequences and currently
only supports the computation of 
the distribution of the number of hits when both
DNA strands are scanned for motif hits.
}
\examples{


seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)
motiffile=system.file("extdata","x31.tab", package="motifcounter")
alpha=0.001
motifcounterOption(alpha)

# estimate a background model
bg=readBackground(seqs,1)

# load a motif from motiffile
motif=t(as.matrix(read.table(motiffile)))

# Compute overlap probabilities
op=motifcounter:::probOverlapHit(motif,bg,singlestranded=FALSE)

# 2 sequences of length 100 bp
seqlen=rep(100,2) 

# Computes the distribution of the number of motif hits
dist=motifcounter:::combinatorialDist(seqlen, op)

}
\seealso{
\code{\link{compoundPoissonDist}}

\code{\link{numMotifHits}}

\code{\link{probOverlapHit}}
}

