% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial_wrapper.R
\name{combinatorialDist}
\alias{combinatorialDist}
\title{Combinatrial model approximation of the number of motif hits}
\usage{
combinatorialDist(seqlen, overlap)
}
\arguments{
\item{seqlen}{Integer-valued vector that defines the lengths of the
individual sequences. This information can be extracted for
given DNA sequence of interest using \code{\link{numMotifHits}}.}

\item{overlap}{Overlap-object that contains the overlapping 
hit probabilities. An Overlap-object is created by 
\code{\link{probOverlapHit}}}
}
\value{
List that contains
\describe{
\item{dist}{Distribution of the number of motif hits}
}
}
\description{
This function approxmiates the distribution of the number of motif hits.
It sums over all combinations of obtaining k hits in a random sequence
of a given length using and efficient dynamic programming algorithm.
This function is an alternative to \code{\link{compoundPoissonDist}}.
However, it requires fixed-length sequences and currently
only supports the distribution of the number of hits when both
DNA strands are scanned for motif hits.
}
\examples{


seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)
motiffile=system.file("extdata","x31.tab", package="motifcounter")
alpha=0.001
motifcounterOption(alpha)

# estimate a background model
bg=readBackground(seqs,1)

# load a motif from motiffile
motif=t(as.matrix(read.table(motiffile)))

# Compute overlap probabilities
op=probOverlapHit(motif,bg,singlestranded=FALSE)
seqlen=rep(100,1)
# Computes the distribution of the number of motif hits
dist=combinatorialDist(seqlen, op)

}
\seealso{
\code{\link{compoundPoissonDist}}
}

