% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial_wrapper.R
\name{combinatorialDist}
\alias{combinatorialDist}
\title{Combinatrial model approximation of the number of motif hits}
\usage{
combinatorialDist(seqlen, overlap)
}
\arguments{
\item{seqlen}{Integer-valued vector that defines the lengths of the
individual sequences. This information can be extracted for
given DNA sequence of interest using \code{\link{numMotifHits}}.
Note that as the combinatorial model only supports the analysis
of a set of sequences of equal length, each element `seqlen`
must be the same.}

\item{overlap}{List that contains the overlapping hit probabilities which
are determined by \code{\link{probOverlapHit}}}
}
\value{
List that contains
\describe{
\item{dist}{Distribution of the number of motif hits}
}
}
\description{
This function approxmiates the distribution of the number of motif hits.
It is determined by efficiently
summing over all combinations of obtaining k hits in a given sequence.
This function can be viewed as an alternative analytical approximation
to \code{\link{compoundPoissonDist}}.
Currently, the combinatorial model approximation is only supported
for scanning both strands of the DNA.
}
\examples{


seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)
motiffile=system.file("extdata","x31.tab", package="motifcounter")
alpha=0.001
motifcounterOption(alpha)

# estimate background model from seqfile
bg=readBackground(seqs,1)

# load motif model from motiffile
motif=t(as.matrix(read.table(motiffile)))

# Compute overlap probabilities
op=probOverlapHit(motif,bg,singlestranded=FALSE)
seqlen=rep(100,1)
# Computes the distribution of the number of motif hits
dist=combinatorialDist(seqlen, op)

}
\seealso{
\code{\link{compoundPoissonDist}}
}

