% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial_wrapper.R
\name{combinatorialDist}
\alias{combinatorialDist}
\title{Combinatrial distribution for the number of motif hits}
\usage{
combinatorialDist(seqlen, overlap, maxhits = 100)
}
\arguments{
\item{seqlen}{Is an integer-valued vector that defines the lengths of the
individual sequences.}

\item{overlap}{A list that contains the overlapping hit probabilities which
is produced by `probOverlapHit()`}

\item{maxhits}{Integer that sets the maximal number of motif hits.}
}
\value{
The distribution of the number of motif hits
}
\description{
This function approxmiates the distribution of the number of motif hits by
summing over all combinations of obtaining k hits in a given sequence. The
summation is done by employing dynamic programming.
}
\examples{


library(mdist)
seqfile=system.file("extdata","seq.fasta", package="mdist")
pwmfile=system.file("extdata","x31.tab", package="mdist")
alpha=0.001
mdistOption(alpha)

# estimate background model from seqfile
readBackground(seqfile,1)

# load motif model from pwmfile
readMotif(pwmfile, 0.01)

# Compute overlap probabilies
op=probOverlapHit(singlestranded=FALSE)
seqlen=rep(100,1)
# Computes the distribution of the number of motif hits
dist=combinatorialDist(seqlen, op)

}

