% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comppoiss_wrapper.R
\name{compoundPoissonDist}
\alias{compoundPoissonDist}
\title{Compound Poisson Approximation}
\usage{
compoundPoissonDist(seqlen, overlap, method = "kopp")
}
\arguments{
\item{seqlen}{Integer-valued vector that defines the lengths of the
individual sequences. This information can be extracted for
given DNA sequence of interest using \code{\link{numMotifHits}} or
\code{\link{lenSequences}}. 
In contrast to \code{\link{combinatorialDist}}, this function
supports variable-length DNA sequence.}

\item{overlap}{Overlap-object that contains the overlapping 
hit probabilities. An Overlap-object is created by 
\code{\link{probOverlapHit}}}

\item{method}{String that defines which method shall be invoked: 'pape' or
'kopp' (see description). Default: method='kopp'}
}
\value{
List containing
\describe{
\item{dist}{Distribution of the number of hits}
}
}
\description{
This function computes the distribution of the number of motif hits
that would be observes in a random DNA sequence of a given length.
The distribution can be determine in two alternative ways:
1) We provide a re-implemented version of the algorithm that was
described in Pape et al. \emph{Compound poisson approximation
of the number of occurrences of a position
frequency matrix (PFM) on both strands.} 2008. 
The main purpose of this method was for benchmarking the improved 
approximation.
In contrast to the original model, this implementation 
can be used with higher-order
background models. To invoke this method use method='pape'.
2) We provide an improved compound Poisson approximation that
uses more appropriate statistical assumptions concerning
overlapping motif hits and that can be used with higher-order
background models as well. The improved version is used by default
with method='kopp'.
}
\examples{


seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)
motiffile=system.file("extdata","x31.tab", package="motifcounter")
alpha=0.001
gran=0.1
motifcounterOption(alpha, gran)

# estimate a background model
bg=readBackground(seqs,1)

# load motif model from motiffile
motif=t(as.matrix(read.table(motiffile)))

# compute the distribution for scanning a single DNA strand
#Compute overlapping probabilities
op=probOverlapHit(motif,bg,singlestranded=TRUE)

# Computes the distribution of the number of motif hits
seqlen=rep(150,100)
dist=compoundPoissonDist(seqlen, op)
#plot(1:length(dist$dist)-1, dist$dist)

# compute the distribution for scanning both DNA strands
#Compute overlapping probabilities
op=probOverlapHit(motif,bg,singlestranded=FALSE)

# Computes the distribution of the number of motif hits
seqlen=rep(150,100)
dist=compoundPoissonDist(seqlen, op)
#plot(1:length(dist$dist)-1, dist$dist)

}
\seealso{
\code{\link{combinatorialDist}}
}

