% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comppoiss_wrapper.R
\name{compoundPoissonDist}
\alias{compoundPoissonDist}
\title{Compound Poisson Approximation}
\usage{
compoundPoissonDist(seqlen, overlap, method = "kopp")
}
\arguments{
\item{seqlen}{Vector of integers that contains the individual sequence
lengths of a given set of sequences. This information can be extracted for 
given DNA sequence of interest using \code{\link{numMotifHits}}.
In contrast to \code{\link{combinatorialDist}}, this function
supports variable length sequence to be considered.}

\item{overlap}{Overlapping hit probabilities. This argument is the result of
\code{link{probOverlapHit}}, which is essential for both.}

\item{method}{String that defines which method shall be invoked 'pape' or
'kopp' (see description). Default: method='kopp'}
}
\value{
List containing
\describe{
\item{dist}{Distribution of the number of hits}
}
}
\description{
This function computes the distribution of the number of motif hits 
that would be observe in a theoretical sequence of length 'L'
that is generated by the background model.
The distribution can be determine in two alternative ways: 
1) We provide a reimplemented version of the algorithm that was
described in Pape et al. \emph{Compound poisson approximation 
of the number of occurrences of a position 
frequency matrix (PFM) on both strands.} 2008. In contrast to the
original model, this implementation can be used with higher-order
background models. This version
is used with method='pape'.
2) We provide an improved compound Poisson approximation that
uses more appropriate statistical assumptions concerning 
overlapping motif hits and that can be used with higher-order
background models as well. The improved version can be used
with method='kopp'.
}
\examples{


seqfile=system.file("extdata","seq.fasta", package="mdist")
motiffile=system.file("extdata","x31.tab", package="mdist")
alpha=0.001
gran=0.1
mdistOption(alpha, gran)

# estimate background model from seqfile
readBackground(seqfile,1)

# load motif model from motiffile
readMotif(motiffile, 0.01)

# compute the distribution for scanning a single DNA strand

#Compute overlapping probabilities
op=probOverlapHit(singlestranded=TRUE)

# Computes the distribution of the number of motif hits
seqlen=rep(150,100)
dist=compoundPoissonDist(seqlen, op)
#plot(1:length(dist$dist)-1, dist$dist)

# compute the distribution for scanning both DNA strands

#Compute overlapping probabilities
op=probOverlapHit(singlestranded=FALSE)

# Computes the distribution of the number of motif hits
seqlen=rep(150,100)
dist=compoundPoissonDist(seqlen, op)
#plot(1:length(dist$dist)-1, dist$dist)

}
\seealso{
\code{\link{combinatorialDist}}
}

