% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{deleteBackground}
\alias{deleteBackground}
\title{Delete background model}
\usage{
deleteBackground()
}
\value{
None
}
\description{
This function unloads the current background model and frees its allocated
memory.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
readBackground(file=seqfile,1)
deleteBackground()


}

