% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forground_wrapper.R
\name{deleteMotif}
\alias{deleteMotif}
\title{Deletes the current motif}
\usage{
deleteMotif()
}
\value{
None
}
\description{
This function unloads the current motif and frees the allocated memory
}
\examples{


motiffile=system.file("extdata","x31.tab", package="mdist")
readMotif(motiffile,0.01)
deleteMotif()

}

