% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{fetchBackground}
\alias{fetchBackground}
\title{Fetch background model into R vectors}
\usage{
fetchBackground()
}
\value{
A list containing the stationary
and the transition probabilities, respectively:
\describe{
\item{stat}{Stationary distribution}
\item{trans}{Transition probabilities}
}
}
\description{
This function returns the parameters of the current background model
as an R list.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

file=system.file("extdata","seq.fasta", package="mdist")
readBackground(file,1)
fetchBackground()

}

