% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_wrapper.R
\name{generateDNAString}
\alias{generateDNAString}
\title{Generate DNAString}
\usage{
generateDNAString(len, bg)
}
\arguments{
\item{len}{Length of the sequence}

\item{bg}{A Background object}
}
\value{
DNAString
}
\description{
This function generates a DNAString
by sampling a sequence from the the background model.
}
\examples{

# Set the the significance level and the score granularity
motifcounterOption(alpha=0.01, gran=0.1)

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)

# Load an order-1 background model
bg=readBackground(seqs,1)

# generates a sequences of length 1kb
generateDNAString(1000,bg)

}

