% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_wrapper.R
\name{generateDNAStringSet}
\alias{generateDNAStringSet}
\title{Generate DNAStringSet}
\usage{
generateDNAStringSet(len)
}
\arguments{
\item{len}{Vector of individual sequence lengths}
}
\value{
DNAStringSet
}
\description{
This function generates a DNAStringSet
by sampling from the background model.
}
\examples{

# Set the the significance level and the score granularity
motifcounterOption(alpha=0.01, gran=0.1)

seqfile=system.file("extdata","seq.fasta", package="motifcounter")

# Load an order-1 background model
readBackgroundForSampling(seqfile,1)

# generates sequences of various lengths
generateDNAStringSet(10:50)

}
\seealso{
\code{\link{readBackgroundForSampling}}
}

