% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_wrapper.R
\name{lenSequences}
\alias{lenSequences}
\title{Length of sequences in a given fasta file}
\usage{
lenSequences(seqs)
}
\arguments{
\item{seqs}{A DNAStringSet object}
}
\value{
A vector containing the lengths of each individual sequences
}
\description{
The function returns a vector containing the lengths
of each individual sequence contained in seqs. Sequences
containing 'N' or 'n' are skipped from the analysis and are
set to length zero.
}
\examples{
file=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(file)
motifcounter:::lenSequences(seqs)

}

