% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{mdistOption}
\alias{mdistOption}
\title{Set parameters for the enrichment analysis}
\usage{
mdistOption(alpha, gran = 0.1, ncores = 1)
}
\arguments{
\item{alpha}{Numerical value that defines the significance level for calling
motif hits. E.g. alpha=0.001 amounts to calling 1 in 1000 nucleotides a
motif hit by chance if a single DNA strand is scanned.}

\item{gran}{Numerical value that defines the granularity which shall be used
for discretizing the score range. Smaller values of gran might increase the
accuracy of the distribution at the cost of an increased computational
runtime. Default: gran=0.1}

\item{ncores}{Number of cores used for parallel processing. Default:
ncores=1.}
}
\description{
This function sets parameters for the enrichment analysis.
}
\examples{


library(mdist)
mdistOption(alpha=0.001)

}

