% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forground_wrapper.R
\name{motif2matrix}
\alias{motif2matrix}
\title{Returns the current PFM as an R matrix}
\usage{
motif2matrix()
}
\value{
A position frequency matrix
}
\description{
This function returns the current PFM as an R matrix.
}
\examples{


motiffile=system.file("extdata","x31.tab", package="mdist")
readMotif(motiffile)
x=motif2matrix()

}

