% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifcounter-package.R
\docType{package}
\name{motifcounter-package}
\alias{motifcounter}
\alias{motifcounter-package}
\title{Motif hit enrichment analysis of DNA sequences}
\description{
The package provides functions for determining the motif hit
enrichment for a given position frequency matrix (PFM)
in a DNA sequence of interest.
To this end, the user is required to provide 1) a PFM (see
\code{link{readMotif}}),
2) a DNA sequence which is used to create
the sequence background model (see \code{link{readBackground}}),
and a desired false positive probability of motif hits to occur in
random sequences generated by the background
(see \code{\link{motifcounterOption}}.
Based on the PFM and the background model, we analyticly determine
the log-likelihood or score distribution, which is used to determine
a score threshold that is associated with the desired
false positive probability.
}
\details{
A motif hit is defined as a position in the sequence at which
the observed score is greater or equal than the score threshold.
We determine the theoretical distribution of
the number of motif hits that emerge
from scanning random DNA sequences that are generated by
the background model in two alternative ways:
1) By a compound Poisson approximation and 2)  by a novel combinatorial
model.
This example guides you through the main functions that are provided
by the package.
}
\examples{


# example input files
seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)
seqs=seqs[1:10]
motiffile=system.file("extdata","x31.tab", package="motifcounter")

# set motif scan options
# alpha corresponds to the desired false positive probability
# of obtaining a motif hits in a random DNA sequence and is typically
# chosen to be 0.01 or lower.
alpha=0.01
motifcounterOption(alpha)

seqlen=lenSequences(seqs)

# Estimate background model as an order-1 Markov model from seqfile
bg=readBackground(seqs,1)

# Load a PFM model from motiffile and add pseudo-evidence of 0.01
# to each element to prevent zero-entry occurrences
motif=t(as.matrix(read.table(motiffile)))

# Score distribution is determined analytically using  dynamic programming
scoredist=scoreDist(motif,bg)
plot(scoredist$score,scoredist$probability, col="green")


#####
# Compute the overlap probabilities with respect to scanning both
# DNA strands. They are necessary for the
# distribution of the number of motif hits and the enrichment test
op=probOverlapHit(motif,bg,singlestranded=FALSE)

# Compound poisson approximation of the motif hits distribution
# in random sequences
cpdist=compoundPoissonDist(seqlen, op)

# combinatorial approximation of the motif hits distribution
# in random sequences
combdist=combinatorialDist(seqlen, op)

#####
# Compute the overlap probabilities with respect to scanning both
# DNA strands. They are necessary for the
# distribution of the number of motif hits and the enrichment test
op=probOverlapHit(motif,bg,singlestranded=TRUE)

# Compound poisson approximation of the motif hits distribution
# in random sequences
cpdist=compoundPoissonDist(seqlen, op)

# Determine the motif enrichment in a given DNA sequence

# Obtain the number of motif hits in the given sequence
numhits=numMotifHits(seqs,motif,bg,singlestranded=TRUE)
op=probOverlapHit(motif,bg, singlestranded=TRUE)
pvalue=motifEnrichmentTest(seqs,motif,bg,
            singlestranded=TRUE,method="compound")

numhits=numMotifHits(seqs,motif,bg,singlestranded=FALSE)
op=probOverlapHit(motif,bg,singlestranded=FALSE)
pvalue=motifEnrichmentTest(seqs,motif,bg,
            singlestranded=FALSE,method="compound")

numhits=numMotifHits(seqs,motif,bg,singlestranded=FALSE)
op=probOverlapHit(motif,bg,singlestranded=FALSE)
pvalue=motifEnrichmentTest(seqs,motif,bg,singlestranded=FALSE,
method="combinatorial")


}
\author{
Wolfgang Kopp

Maintainer: Wolfgang Kopp <kopp@molgen.mpg.de>
}
\seealso{
\code{\link{compoundPoissonDist}}, \code{\link{combinatorialDist}}
\tabular{ll}{ Package: \tab motifcounter\cr
Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2016-11-04\cr License: \tab GPL-2\cr }
}
\keyword{MotifEnrichment}
\keyword{PFM,}

