% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentTest.R
\docType{methods}
\name{motifEnrichment}
\alias{motifEnrichment}
\alias{motifEnrichment,DNAStringSet,matrix-method}
\alias{motifEnrichment,DNAStringSetList,matrix-method}
\alias{motifEnrichment,GRanges,matrix-method}
\alias{motifEnrichment,GRangesList,matrix-method}
\alias{motifEnrichment,character,matrix-method}
\alias{motifEnrichment,list,matrix-method}
\alias{motifEnrichment,ANY,list-method}
\title{Enrichment of motif hits}
\usage{
motifEnrichment(seqs_or_regions, pfms, ...)

\S4method{motifEnrichment}{DNAStringSet,matrix}(seqs_or_regions, pfms, bg,
  genome = NULL, singlestranded = FALSE, method = "compound",
  normalize_pfm = TRUE)

\S4method{motifEnrichment}{DNAStringSetList,matrix}(seqs_or_regions, pfms,
  bg, genome = NULL, singlestranded = FALSE, method = "compound",
  normalize_pfm = TRUE)

\S4method{motifEnrichment}{GRanges,matrix}(seqs_or_regions, pfms, bg,
  genome = NULL, singlestranded = FALSE, method = "compound",
  normalize_pfm = TRUE)

\S4method{motifEnrichment}{GRangesList,matrix}(seqs_or_regions, pfms, bg,
  genome = NULL, singlestranded = FALSE, method = "compound",
  normalize_pfm = TRUE)

\S4method{motifEnrichment}{character,matrix}(seqs_or_regions, pfms, bg,
  genome = NULL, singlestranded = FALSE, method = "compound",
  normalize_pfm = TRUE)

\S4method{motifEnrichment}{list,matrix}(seqs_or_regions, pfms, bg,
  genome = NULL, singlestranded = FALSE, method = "compound",
  normalize_pfm = TRUE)

\S4method{motifEnrichment}{ANY,list}(seqs_or_regions, pfms, bg,
  genome = NULL, singlestranded = FALSE, method = "compound",
  normalize_pfm = TRUE)
}
\arguments{
\item{seqs_or_regions}{DNAStringSet, DNAStringSetList,
GRanges, GRangesList,
character-string denoting bed-file location or
list(character) denoting a list of bed-files}

\item{pfms}{matrix or list of matrices representing position weight matrices}

\item{...}{see concrete methods}

\item{bg}{A Background object}

\item{genome}{BSgenome object from which the sequences are extracted
using the provided regions. Only used in conjunction with
GRanges, GRangesList, and regions extracted from bed-files.}

\item{singlestranded}{Boolean that indicates whether a single strand or
both strands shall be scanned for motif hits.
Default: singlestranded = FALSE.}

\item{method}{String that defines whether to use
the 'compound' Poisson approximation' or the 'combinatorial' model.
Default: method='compound'.}

\item{normalize_pfm}{If TRUE, normalization of the PFM is performed
to avoid zero-entries. Default: FALSE}
}
\value{
List that contains
\describe{
\item{pvalue}{P-value for the enrichment test.}
\item{fold}{Fold-enrichment with respect to the expected number of hits}
\item{logfold}{log-Fold-enrichment with respect to
               the expected number of hits}
}
}
\description{
This function determines whether a given motif is enriched in a given
DNA sequences.
}
\details{
Enrichment is tested by comparing the observed
number of motif hits against a theoretical distribution of the number
of motif hits in random DNA sequences.
Optionally, the theoretical distribution of the number of motif
hits can be evaluated by either a 'compound Poisson model'
or the 'combinatorial model'.
Additionally, the enrichment test can be conducted with respect
to scanning only the forward strand or both strands of the DNA
sequences. The latter option is only available for the
'compound Poisson model'.
}
\examples{


# Load sequences
seqfile = system.file("extdata", "seq.fasta", package = "motifcounter")
seqs = Biostrings::readDNAStringSet(seqfile)

# Load background
bg = readBackground(seqs, 1)

# Load motif
motiffile = system.file("extdata", "x31.tab", package = "motifcounter")
motif = t(as.matrix(read.table(motiffile)))

# 1 ) Motif enrichment test w.r.t. scanning a *single* DNA strand
# based on the 'Compound Poisson model'

result = motifEnrichment(seqs, motif, bg,
            singlestranded = TRUE, method = "compound")

# 2 ) Motif enrichment test w.r.t. scanning *both* DNA strand
# based on the 'Compound Poisson model'

result = motifEnrichment(seqs, motif, bg, method = "compound")

# 3 ) Motif enrichment test w.r.t. scanning *both* DNA strand
# based on the *combinatorial model*

result = motifEnrichment(seqs, motif, bg, singlestranded = FALSE,
            method = "combinatorial")

# 4 ) Evaluate the enrichment in two sets of sequences
#     (e.g. promoters and enhancers) using three motifs.

result = motifEnrichment(Biostrings::DNAStringSetList(seqs, seqs),
                         list(motif, motif, motif),
                         bg)

# 4 ) Evaluate the enrichment in two sets of genomic regions
#     exracted from the human genome
#     (e.g. promoters and enhancers) using three motifs

regions1 = system.file("extdata",
                       "example_jund.bed", package = "motifcounter")
regions2 = system.file("extdata",
                       "example_arid3a.bed", package = "motifcounter")
regions = GenomicRanges::GRangesList(
           "jund"=genomation::readBed(regions1),
           "arid3a"=genomation::readBed(regions2))

genome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38

query <- MotifDb::query
motifs1 = as.list(query(query(MotifDb::MotifDb, "hsapiens"), c("arid3a")))
motifs2 = as.list(query(query(MotifDb::MotifDb, "hsapiens"), c("jund")))
motifs = c(motifs1[1], motifs2[1])

result = motifEnrichment(regions, motifs, bg, genome=genome)

}
\seealso{
\code{\link{compoundPoissonDist}}, \code{\link{combinatorialDist}}
}
