% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentTest.R
\name{motifEnrichment}
\alias{motifEnrichment}
\title{Enrichment of motif hits}
\usage{
motifEnrichment(seqs, pfm, bg, singlestranded = FALSE, method = "compound")
}
\arguments{
\item{seqs}{DNAString or DNAStringSet}

\item{pfm}{A position frequency matrix}

\item{bg}{A Background object}

\item{singlestranded}{Boolean flag that indicates whether one or both
strands are scanned for motif hits}

\item{method}{String that defines whether to use
the 'compound' Poisson approximation' or the 'combinatorial' model.
Default: method='compound'}
}
\value{
Result list that contains
\described{
\item{pvalue}{P-value for the enrichment test}
\item{fold}{Fold-enrichment with respect to the expected number of hits}
}
}
\description{
This function determines whether a given motif is enriched in a given
DNA sequences based on the number of motif hits that would be observed
in a random sequence that is generated by the background model.
It can optionally be used based on the
'compound Poisson approximation' or the 'combinatorial model'.
}
\examples{


seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)
motiffile=system.file("extdata","x31.tab", package="motifcounter")
alpha=0.001
gran=0.1
motifcounterOption(alpha, gran)

# estimate background model from seqfile
bg=readBackground(seqs,1)

# load motif model from motiffile
motif=t(as.matrix(read.table(motiffile)))

### 1 ) Compute the distribution for scanning a *single* DNA strand
# based on the 'Compound Poisson model'

result=motifEnrichment(seqs,motif,bg,
            singlestranded=TRUE,method="compound")

### 2 ) Compute the distribution for scanning *both* DNA strand
# based on the 'Compound Poisson model'

result=motifEnrichment(seqs,motif, bg, method="compound")

### 3 ) Compute the distribution for scanning *both* DNA strand
# based on the *combinatorial model*

result=motifEnrichment(seqs,motif, bg,singlestranded=FALSE,
            method="combinatorial")

}
\seealso{
\code{\link{compoundPoissonDist}}, \code{\link{combinatorialDist}}
}

