% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentTest.R
\name{motifEnrichmentTest}
\alias{motifEnrichmentTest}
\title{Enrichment of motif hits}
\usage{
motifEnrichmentTest(sequence, singlestranded = FALSE, method = "compound")
}
\arguments{
\item{sequence}{Fasta-filename containing the sequence that should
be scanned}

\item{singlestranded}{Boolean flag that indicates whether one or both
strands are scanned for motif hits}

\item{method}{String that defines whether to use 
the 'compound' Poisson approximation' or the 'combinatorial' model.
Default: method='compound'}
}
\value{
P-value for the enrichment test
}
\description{
This function determines whether a given motif is enriched in a given 
DNA sequences based on the number of motif hits that would be observed
in a random sequence that is generated by the background model. 
It can optionally be used based on the 
'compound Poisson approximation' or the 'combinatorial model'.
}
\examples{


seqfile=system.file("extdata","seq.fasta", package="motifcounter")
motiffile=system.file("extdata","x31.tab", package="motifcounter")
alpha=0.001
gran=0.1
motifcounterOption(alpha, gran)

# estimate background model from seqfile
readBackground(seqfile,1)

# load motif model from motiffile
readMotif(motiffile, 0.01)

### 1 ) Compute the distribution for scanning a *single* DNA strand
# based on the 'Compound Poisson model'

pvalue=motifEnrichmentTest(seqfile,singlestranded=TRUE,method="compound")

### 2 ) Compute the distribution for scanning *both* DNA strand
# based on the 'Compound Poisson model'

pvalue=motifEnrichmentTest(seqfile,method="compound")

### 3 ) Compute the distribution for scanning *both* DNA strand
# based on the *combinatorial model*

pvalue=motifEnrichmentTest(seqfile,singlestranded=FALSE,
method="combinatorial")

}
\seealso{
\code{\link{compoundPoissonDist}}, \code{\link{combinatorialDist}}
}

