% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_wrapper.R
\name{motifHitProfile}
\alias{motifHitProfile}
\title{Motif hit profile across multiple sequences}
\usage{
motifHitProfile(seqs, pfm, bg)
}
\arguments{
\item{seqs}{DNAStringSet}

\item{pfm}{A position frequency matrix}

\item{bg}{A Background object}
}
\value{
List containing
\describe{
\item{fscores}{Average forward strand motif hits}
\item{rscores}{Average reverse strand motif hits}
}
}
\description{
This function computes the average motif hit profile across a set of DNA sequences.
}
\examples{



# Load the DNA sequence and a Motif
seqfile=system.file("extdata","oct4_chipseq.fa", package="motifcounter")
motiffile=system.file("extdata","x31.tab",package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)
seqs=seqs[1:10]

# Load the order-1 background model from the DNA sequence
bg=readBackground(seqs,1)

# Load the motif from the motiffile
motif=t(as.matrix(read.table(motiffile)))

# Compute the score distribution
motifHitProfile(seqs,motif,bg)

}

