% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forground_wrapper.R
\name{motifLength}
\alias{motifLength}
\title{Length of the currently loaded PFM}
\usage{
motifLength()
}
\description{
This function returns the length of the currently loaded PFM.
}
\examples{


library(mdist)
pwmfile=system.file("extdata","x31.tab", package="mdist")
readMotif(pwmfile)
motifLength()

}

