% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_wrapper.R
\name{numMotifHits}
\alias{numMotifHits}
\title{Number of motif hits in a given DNA sequence}
\usage{
numMotifHits(seqfile, singlestranded = FALSE)
}
\arguments{
\item{seqfile}{Name of the sequence file in fasta format.}

\item{singlestranded}{Boolian flag that indicates whether a single strand or
both strands shall be scanned for motif hits}
}
\description{
The function scans a given DNA sequence or a set of sequences and counts the
number of motif hits on both, the forward and the reverse strand. It returns
the length of each sequence supplied in the fasta file and the number of
hits observed across all sequences.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

library(mdist)
seqfile=system.file("extdata","seq.fasta", package="mdist")
pwmfile=system.file("extdata","x31.tab", package="mdist")
alpha=0.001
mdistOption(alpha)

readBackground(seqfile,1)
readMotif(pwmfile,0.01)

# scan the given sequence on both strands for the motif occurances
noc=numMotifHits(seqfile)
noc

# scan the given sequence on a single strand for the motif occurances
noc=numMotifHits(seqfile,singlestranded=TRUE)
noc

}

