% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_wrapper.R
\name{numSequences}
\alias{numSequences}
\title{Number of sequences in a given fasta file}
\usage{
numSequences(seqs)
}
\arguments{
\item{seqs}{DNAStringSet}
}
\value{
The number of sequences
}
\description{
The function parses a fasta file and returns the number of individual
sequence that it contains.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)

# scan the given sequence for the motif occurances
nseq=numSequences(seqs)
nseq

}

