% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_wrapper.R
\name{numSequences}
\alias{numSequences}
\title{Number of sequences in a given fasta file}
\usage{
numSequences(seqs)
}
\arguments{
\item{seqs}{DNAStringSet}
}
\value{
The number of sequences
}
\description{
The function returns the number of individual
sequence.
}

