% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{printBackground}
\alias{printBackground}
\title{Prints the current background model}
\usage{
printBackground()
}
\description{
This function prints the currently loaded background model.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

library(mdist)
seqfile=system.file("extdata","seq.fasta", package="mdist")
readBackground(file=seqfile,1)
printBackground()

}

