% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{printBackgroundForSampling}
\alias{printBackgroundForSampling}
\title{Prints the current background model for sampling}
\usage{
printBackgroundForSampling()
}
\description{
This function prints the currently loaded background model that is employed
during scanning the sampled DNA sequence. It is not used for the compound
Poisson model and the combinatorial model
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

library(mdist)
seqfile=system.file("extdata","seq.fasta", package="mdist")
readBackgroundForSampling(file=seqfile,1)
printBackgroundForSampling()


}

