% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{printBackgroundForSampling}
\alias{printBackgroundForSampling}
\title{Prints the current background model for sampling}
\usage{
printBackgroundForSampling()
}
\value{
None
}
\description{
Similar to \code{link{printBackgroundForSampling}}, but
prints parameters that where acquired using
\code{link{readBackgroundForSampling}}.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
readBackgroundForSampling(file=seqfile,1)
printBackgroundForSampling()


}
\seealso{
\code{link{printBackgroundForSampling}}
}

