% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comppoiss_wrapper.R
\name{probOverlapHit}
\alias{probOverlapHit}
\title{Motif overlap probabilities}
\usage{
probOverlapHit(singlestranded = FALSE)
}
\arguments{
\item{singlestranded}{Boolian which defines whether the overlapping hit
probabilities shall be computed with respect to scanning both DNA strands or
only one strand.  Default: Both strands are scanned for DNA sequences.}
}
\value{
A list containing various overlapping hit probabilities,  including
            beta ... the overlapping hit probability for hits occurring on the same strand
}
\description{
This function computes the self-overlapping probabilites of the motif for
the loaded PWM and background based on dynamic programming.
}
\examples{


library(mdist)
seqfile=system.file("extdata","seq.fasta", package="mdist")
pwmfile=system.file("extdata","x31.tab", package="mdist")
alpha=0.001
gran=0.01
mdistOption(alpha, gran)

# estimate background model from seqfile
readBackground(seqfile,1)

# load motif model from pwmfile
readMotif(pwmfile,0.01)

# compute the overlap probabilities for scanning both DNA strands
op=probOverlapHit(singlestranded=FALSE)

# compute the overlap probabilities for scanning a single DNA strand
op=probOverlapHit(singlestranded=TRUE)

}

