% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{readBackground}
\alias{readBackground}
\title{Estimates the background model}
\usage{
readBackground(file, order = 1)
}
\arguments{
\item{file}{Fasta-filename.}

\item{order}{Order of the Markov models that shall be used as the
background model. Default: order=1.}
}
\value{
None
}
\description{
This function reads a DNA sequence from a given fasta file
and uses that sequence to estimate an order-m Markov model.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

file=system.file("extdata","seq.fasta", package="motifcounter")
readBackground(file,1)

}

