% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{readBackground}
\alias{readBackground}
\title{Estimates the background model}
\usage{
readBackground(file, order = 1)
}
\arguments{
\item{file}{Filename where the sequence is stored in. The file must be in
fasta format.}

\item{order}{Defines the order of the Markov models used for the background.
Default: order=1.}
}
\description{
This function reads a DNA sequence in fasta format and estimates and order m
Markov model.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

library(mdist)
file=system.file("extdata","seq.fasta", package="mdist")
readBackground(file,1)

}

