% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_wrapper.R
\name{readBackgroundForSampling}
\alias{readBackgroundForSampling}
\title{Estimates the background model for sampling}
\usage{
readBackgroundForSampling(file, order = 1)
}
\arguments{
\item{file}{Name of the sequence file in fasta format.}

\item{order}{Defines the order of the Markov models used for the background.
Default: order=1.}
}
\description{
This function reads a DNA sequence in fasta format and estimates and order m
Markov model. This function is only used when simulated DNA sequences are
scanned by the model.  The compound Poisson model and the combinatorial
model do not use this backgroud model.
}
\examples{

# Estimate first order Markov model based on the sequence provided
# in seq.fasta

library(mdist)
file=system.file("extdata","seq.fasta", package="mdist")
readBackgroundForSampling(file,1)


}

