% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forground_wrapper.R
\name{readMotif}
\alias{readMotif}
\title{Read and load a PFM}
\usage{
readMotif(motif, pseudocount = 0.01)
}
\arguments{
\item{motif}{Filename to load the PFM from
or PFM provided as an R matrix.}

\item{pseudocount}{Numeric value which is added to each element of the
PFM in order to avoid zero probability elements. The columns of the
matrix are renormalized automatically.}
}
\value{
None
}
\description{
The function reads a position frequency matrix (PFM) for scanning a DNA
sequence. The PFM can be loaded directly in transfac format, in tab format
or from an R matrix. To load a PFM from a file directly, the files must end
in '.transfac' or '.tab' to load from transfac or tab format, respectively,
and each file must contain only one PFM.
}
\examples{


motiffile=system.file("extdata","x31.tab", package="motifcounter")
# load a motif in tab format
readMotif(motiffile, 0.01)

# fetch the motif to an R matrix
motif=motif2matrix()

#reload the motif again from the R matrix
readMotif(motif, 0.01)


}

