% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_wrapper.R
\name{scoreDist}
\alias{scoreDist}
\title{Score distribution}
\usage{
scoreDist()
}
\value{
List containing 
\describe{
\item{score}{Vector of motif scores}
\item{probability}{Score probabilities}
}
}
\description{
This function computes the score distribution for the provided PFM and 
background. The Score distribution is computed based on dynamic
programming.
}
\examples{


# Set the significance level and granularity for the score computation
motifcounterOption(alpha=0.01,gran=0.1)

# Load the DNA sequence and a Motif
seqfile=system.file("extdata","seq.fasta", package="motifcounter")
motiffile=system.file("extdata","x31.tab",package="motifcounter")

# Load the order-1 background model from the DNA sequence
readBackground(seqfile,1)

# Load the motif from the motiffile
readMotif(motiffile)

# Compute the score distribution
dp=scoreDist()

}

