% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_wrapper.R
\name{scoreDistBf}
\alias{scoreDistBf}
\title{Score distribution}
\usage{
scoreDistBf(pfm)
}
\arguments{
\item{pfm}{A position frequency matrix}
}
\value{
List containing
\describe{
\item{score}{Vector of motif scores}
\item{probability}{Score probabilities}
}
}
\description{
This function computes the score distribution for the provided PFM and
background model. The result is identical to \code{\link{scoreDist}},
however, the method employs a less efficient algorithm that
enumerates all DNA sequences of the length of the motif.
This function is only used for debugging purposes
and might require substantial computational
resources for long motifs. Therefore, use \code{\link{scoreDist}} instead.
}
\examples{


# Set the significance level and granularity for the score computation
motifcounterOption(alpha=0.01,gran=0.1)

# Load the DNA sequence and a Motif
seqfile=system.file("extdata","seq.fasta", package="motifcounter")
motiffile=system.file("extdata","x31.tab",package="motifcounter")

# Load the order-1 background model from the DNA sequence
readBackground(seqfile,1)

# Load the motif from the motiffile
motif=t(as.matrix(read.table(motiffile)))

# Compute the score distribution
dp=scoreDistBf(motif)

}
\seealso{
\code{\link{scoreDist}}
}

