% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_wrapper.R
\name{scoreDistBf}
\alias{scoreDistBf}
\title{Score distribution}
\usage{
scoreDistBf()
}
\description{
This function computes the score distribution for the provided PFM and order
m background based on enumerating all DNA sequences of the 
length of the motif. This function is only used for debugging
\code{\link{scoreDist}}, and might require substantial computational
resources for long motifs.
}
\examples{


library(mdist)
# Set the significance level and granularity for the score computation
mdistOption(alpha=0.01,gran=0.1)

# Load the DNA sequence and a Motif
seqfile=system.file("extdata","seq.fasta", package="mdist")
pwmfile=system.file("extdata","x31.tab",package="mdist")

# Load the order-1 background model from the DNA sequence
readBackground(seqfile,1)

# Load the motif from the pwmfile
readMotif(pwmfile)

# Compute the score distribution
dp=scoreDistBf()

}

