% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_wrapper.R
\name{scoreDistEmpirical}
\alias{scoreDistEmpirical}
\title{Empirical score distribution}
\usage{
scoreDistEmpirical(pfm, seqlen, nsim)
}
\arguments{
\item{pfm}{A position frequency matrix}

\item{seqlen}{Length of the sequence to be
simulated.}

\item{nsim}{Number of samples to be generated.}
}
\value{
List containing
\describe{
\item{score}{Vector of scores}
\item{probability}{Vector of score probabilities}
}
}
\description{
This function estimates the empirical score distribution
by simulating random DNA sequences based on the
background model. Subsequently, the random sequences are scanned
with the scoring measure which yields the score distribution.
}
\examples{

# Set the the significance level and the score granularity
motifcounterOption(alpha=0.01, gran=0.1)

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
motiffile=system.file("extdata","x31.tab",package="motifcounter")

# Load an order-1 background model
readBackground(seqfile,1)

# Load a motif from the motiffile
motif=t(as.matrix(read.table(motiffile)))

# generate the simulated score distribution on
# sequences of length 1kb using 1000 samples
scoreDistEmpirical(motif,seqlen= 1000,nsim=1000)

}
\seealso{
\code{\link{scoreDist}}
}

