% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_wrapper.R
\name{scoreHistogramSingleSeq}
\alias{scoreHistogramSingleSeq}
\title{Score histogram on a single sequence}
\usage{
scoreHistogramSingleSeq(seq, pfm, bg)
}
\arguments{
\item{seq}{DNAString}

\item{pfm}{A position frequency matrix}

\item{bg}{A Background object}
}
\value{
List containing
\describe{
\item{score}{Vector of score bins}
\item{probability}{Empirical frequencies associated with each score}
}
}
\description{
This function computes the score histogram
for a given sequence
}

