% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_wrapper.R
\name{scoreHistogramSingleSeq}
\alias{scoreHistogramSingleSeq}
\title{Score histogram on a single sequence}
\usage{
scoreHistogramSingleSeq(seq, pfm)
}
\arguments{
\item{seq}{DNAString}

\item{pfm}{A position frequency matrix}
}
\value{
List containing
\describe{
\item{score}{Vector of score bins}
\item{probability}{Frequencies}
}
}
\description{
This function computes the observed score histogram
for a given sequence
}
\examples{

# Set the the significance level and the score granularity
motifcounterOption(alpha=0.01, gran=0.1)

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
motiffile=system.file("extdata","x31.tab",package="motifcounter")

# Load an order-1 background model
readBackground(seqfile,1)
readBackgroundForSampling(seqfile,1)

# Load a motif from the motiffile
motif=t(as.matrix(read.table(motiffile)))
seq=generateDNAString(1000)

# generate the simulated score distribution on
# sequences of length 1kb using 1000 samples
scoreHistogram(motif,seq)

}
\seealso{
\code{\link{scoreDist}}
}

