% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_wrapper.R
\name{scoreSequence}
\alias{scoreSequence}
\title{Score observations}
\usage{
scoreSequence(seq, pfm, bg)
}
\arguments{
\item{seq}{DNAString}

\item{pfm}{A position frequency matrix}

\item{bg}{A Background object}
}
\value{
List containing
\describe{
\item{fscores}{Scores on the forward strand}
\item{rscores}{Scores on the reverse strand}
}
}
\description{
This function computes the observed score in a given DNA sequence
}
\examples{


# Set the significance level and granularity for the score computation
motifcounterOption(alpha=0.01,gran=0.1)

# Load the DNA sequence and a Motif
seqfile=system.file("extdata","seq.fasta", package="motifcounter")
motiffile=system.file("extdata","x31.tab",package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)

# Load the order-1 background model from the DNA sequence
bg=readBackground(seqs,1)

# Load the motif from the motiffile
motif=t(as.matrix(read.table(motiffile)))

# Compute the score distribution
scoreSequence(seqs[[1]],motif,bg)

}

