% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_wrapper.R
\name{scoreThreshold}
\alias{scoreThreshold}
\title{Score threshold}
\usage{
scoreThreshold(pfm, bg)
}
\arguments{
\item{pfm}{A position frequency matrix}

\item{bg}{A Background object}
}
\value{
List containing
\describe{
\item{threshold}{Score threshold}
\item{alpha}{False positive probability}
}
}
\description{
This function computes the score threshold for a desired
false positive rate to obtain a motif hit.
The returned false positive rate usually differs slightly
from the one that is set in \code{\link{motifcounterOption}}, because
of the discrete nature of the sequences.
}
\examples{

# Set the the significance level and the score granularity
motifcounterOption(alpha=0.01, gran=0.1)

seqfile=system.file("extdata","seq.fasta", package="motifcounter")
motiffile=system.file("extdata","x31.tab",package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)

# Load an order-1 background model
bg=readBackground(seqs,1)

# Load a motif
motif=t(as.matrix(read.table(motiffile)))

# Compute the score threshold
scoreThreshold(motif,bg)

}
\seealso{
\code{\link{scoreDist}}
}

