% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_wrapper.R
\name{simulateNumHitsDist}
\alias{simulateNumHitsDist}
\title{Empirical number of motif hits distribution}
\usage{
simulateNumHitsDist(pfm, bg, seqlen, nsim, singlestranded = FALSE)
}
\arguments{
\item{pfm}{A position frequency matrix}

\item{bg}{A Background object}

\item{seqlen}{Integer-valued vector contanting the individual
sequence lengths.}

\item{nsim}{Number of random samples.}

\item{singlestranded}{Boolian flag that indicates whether a single strand or
both strands shall be scanned for motif hits.}
}
\value{
Empirical number of motif hits distribution
}
\description{
This function simulates random DNA sequences according to
the background model and
subsequently counts how many motif hits occur in them.
Doing that repeatedly eventually established the empirical
distribution of the number of motif hits.
}
\examples{


motifcounterOption(0.01, 0.01)
seqfile=system.file("extdata","seq.fasta", package="motifcounter")
seqs=Biostrings::readDNAStringSet(seqfile)
motiffile=system.file("extdata","x31.tab",package="motifcounter")

bg=readBackground(seqs,1)
motif=t(as.matrix(read.table(motiffile)))

seqlen=rep(150,1)
simc=simulateNumHitsDist(motif, bg, seqlen, nsim=100,singlestranded=FALSE)

simc=simulateNumHitsDist(motif, bg, seqlen, nsim=100,singlestranded=TRUE)

}
\seealso{
\code{\link{compoundPoissonDist}},\code{\link{combinatorialDist}}
}

