% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_wrapper.R
\name{simulateNumHitsDist}
\alias{simulateNumHitsDist}
\title{Empirical number of motif hits distribution}
\usage{
simulateNumHitsDist(seqlen, maxhits, nsim, singlestranded = FALSE)
}
\arguments{
\item{seqlen}{Integer-valued vector contanting the individual
sequence lengths.}

\item{maxhits}{Maximal number of hits.}

\item{nsim}{Number of random samples.}

\item{singlestranded}{Boolian flag that indicates whether a single strand or
both strands shall be scanned for motif hits.}
}
\value{
Empirical number of motif hits distribution
}
\description{
This function simulates random DNA sequences according to 
the background model and
subsequently counts how many motif hits occur in them.
Doing that repeatedly eventually established the empirical
distribution of the number of motif hits.
}
\examples{


motifcounterOption(0.01, 0.01)
seqfile=system.file("extdata","seq.fasta", package="motifcounter")
motiffile=system.file("extdata","x31.tab",package="motifcounter")

readBackground(seqfile,1)
readMotif(motiffile)

seqlen=rep(150,100)
simc=simulateNumHitsDist(seqlen,maxhits=1000,nsim=100,singlestranded=FALSE)

simc=simulateNumHitsDist(seqlen,maxhits=1000,nsim=100,singlestranded=TRUE)

}
\seealso{
\code{\link{compoundPoissonDist}},\code{\link{combinatorialDist}}
}

