% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentTest.R
\name{topMotifs}
\alias{topMotifs}
\title{Extract top matches per region}
\usage{
topMotifs(menr, top_n = 3, doplot = FALSE)
}
\arguments{
\item{menr}{Motif enrichment results obtained by calling motifEnrichment().}

\item{top_n}{Number of top matching motifs to extract. Default: 3.}

\item{doplot}{Whether to plot the results. Default: FALSE}
}
\value{
result table
}
\description{
This function assembles a concise summary of the enrichment results.
It enables to extract the top matching motifs and visualize them.
}
